Address review: type concurrency as an integer, drop README section
Per review feedback, checked against the Buildkite pipeline schema
(buildkite/pipeline-schema schema.json, definitions/commandStep):

- `concurrency` is `{"type": "integer"}` — so `Step.Concurrency` is now
  `int` rather than `interface{}`, and plugin.yml declares `integer`
  instead of `[string, integer]`. The two tests that asserted on
  `float64` (what `interface{}` used to receive from json.Unmarshal)
  now assert on plain int literals.
- `concurrency_method` is a string constrained to
  `enum: ["ordered", "eager"]`, so plugin.yml now carries that enum.
- `concurrency_group` is `{"type": "string"}`, already correct.

Also removed the `### concurrency` README section — the standard
Buildkite keywords are expected to work, so documenting them again here
is redundant. The keys remain in the Detailed example, which is the
canonical list of supported config attributes.

+++ pipeline_test.go
@@ -393,9 +393,7 @@ func TestGeneratePipeline(t *testing.T) {
 				{GithubStatus: GithubStatusNotification{Context: "my-custom-status"}},
 				{Slack: "@someuser", Condition: "build.state === \"passed\""},
 			},
-			// float64 is what json.Unmarshal produces for a plugin config
-			// `"concurrency": 2`, so assert on that rather than an int literal.
-			Concurrency:       float64(2),
+			Concurrency:       2,
 			ConcurrencyGroup:  "payments/deploy",
 			ConcurrencyMethod: "eager",
 		},
